+++ src/gui/annotator/tabs/AnnotationTabContent.h
@@ -27,7 +27,7 @@
 #include "src/gui/annotator/settings/AnnotationItemSettings.h"
 #include "src/annotations/core/AnnotationArea.h"
 #include "src/common/provider/DevicePixelRatioScaler.h"
-#include "src/gui/annotator/AnnotationViewZoomer.h"
+#include "src/gui/scrollAndZoomView/ViewZoomer.h"
 
 namespace kImageAnnotator {